fixed conversion warning

+++ nn/techniques/sgd.cpp
@@ -4,7 +4,7 @@ namespace nn {
 void SGD::optimize(Hyperparameters hp) {
   for (auto &param : params_) {
     auto decay_term = param->weights.w * hp.weight_decay;
-    param->weights -= hp.learningrate * param->gradient / hp.batch_size;
+    param->weights -= hp.learningrate * param->gradient / (float)hp.batch_size;
     param->weights.w -= decay_term;
     param->gradient.zero();
   }
